requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Background Jobs
apscheduler==3.10.4

# SSH Client
paramiko==3.4.0

# HTTP Client
requests==2.31.0
httpx==0.26.0

# Configuration
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
pyyaml==6.0.1

# Utilities
python-dateutil==2.8.2
pytz==2024.1

# Logging
structlog==24.1.0

# Testing (optional, for development)
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0  # For testing FastAPI

# Type checking (optional, for development)
mypy==1.8.0
types-pyyaml==6.0.12.12
types-requests==2.31.0.20240125
